Added a button to the Main.Storyboard, aligned it using Auto Layout to the center of the screen. Added code in the ViewController that prompts the user to allow notifications when opening the app for the first time. Also added an action function for the button that pushes a notification when the user presses the button. The notification will not appear when the app is open, so there is 10 seconds from when the button is pressed to when the notification appears. Hit the button and close out of the app, the notification will appear in several seconds.
The UNNotification framework used by iOS has a built in trigger function that allows a notifications to be pushed depending on location. The one used here is a time interval trigger function. The location one take lat and long as parameters, and also allows a circular region. This will be very helpful when setting up location notifications in the future.